+++ docs/src/routes.tsx
@@ -8,7 +8,6 @@ import DocsRoot from './components/DocsRoot'
 import Introduction from './views/Introduction'
 import PageNotFound from './views/PageNotFound'
 import QuickStart from './views/QuickStart'
-import Accessibility from './views/Accessibility'
 
 const Router = () => (
   <BrowserRouter basename={__BASENAME__}>
@@ -18,7 +17,13 @@ const Router = () => (
         <DocsLayout exact path="/" component={Introduction} />
         <DocsLayout exact path="/:type/:name" component={DocsRoot} sidebar />
         <DocsLayout exact path="/quick-start" component={QuickStart} />
-        <DocsLayout exact path="/accessibility" component={Accessibility} />
+        {process.env.NODE_ENV !== 'production' && (
+          <DocsLayout
+            exact
+            path="/prototype-chat-pane"
+            component={require('./prototypes/chatPane/index').default}
+          />
+        )}
         <DocsLayout exact path="/*" component={PageNotFound} />
       </Switch>
     </Switch>