+++ src/components/Chat/index.ts
@@ -1,2 +1,2 @@
 export { default } from './Chat'
-export { default as ChatMessage } from './ChatMessage'
+export { default as ChatMessage, IChatMessageProps } from './ChatMessage'